alpha version, todo: fix api failure summary, add unit tests, add integration tests

+++ src/Services/IInputFileService.cs
@@ -4,6 +4,6 @@ namespace TemperatureHistogramChallenge.Services
 {
     public interface IInputFileService
     {
-        IDictionary<double, int> ProcessFile(string input);
+        IDictionary<float, int> ProcessFile(string input);
     }
 }
